requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
google-genai
google-api-python-client
google-auth
